Refactor data handling and set resource encoding.
Updated `Data` class to use integer altitude and fixed JSON conversion accordingly. Adjusted resource encoding in `pom.xml` and `.idea/encodings.xml`. Added `api_checker.py` for API testing with validation tests to ensure consistent data handling.

+++ TP5 Sécurisation de l'accès à l'API REST/sporttrack-model/src/main/java/fr/ubs/sporttrack/model/Data.java
@@ -14,11 +14,11 @@ public class Data {
     private int cardioFrequency;
     private float latitude;
     private float longitude;
-    private float altitude;
+    private int altitude;
     
     public Data(){}
     
-    public Data(String time, int cf, float lat, float lon, float alt){
+    public Data(String time, int cf, float lat, float lon, int alt){
         this.time = time;
         this.cardioFrequency = cf;
         this.latitude = lat;
@@ -47,7 +47,7 @@ public class Data {
     }
 
     public static Data fromJSON(JSONObject obj){
-        return new Data(obj.getString(TIME_FIELD), obj.getInt(CARDIO_FREQ_FIELD), obj.getFloat(LATITUDE_FIELD), obj.getFloat(LONGITUDE_FIELD), obj.getFloat(ALTITUDE_FIELD));
+        return new Data(obj.getString(TIME_FIELD), obj.getInt(CARDIO_FREQ_FIELD), obj.getFloat(LATITUDE_FIELD), obj.getFloat(LONGITUDE_FIELD), obj.getInt(ALTITUDE_FIELD));
     }
 
     public JSONObject toJSON(){
